fifth

+++ src/controllers/users.js
@@ -2,46 +2,38 @@ const User = require("../models/user");
 const jwt = require("jsonwebtoken");
 
 exports.addUser = async (req, res) => {
-  const temp = {
-    name: req.body.name,
-    email: req.body.email,
-    password: req.body.password,
-  };
+  const { name, email, password } = req.body;
 
   try {
-    let user = await User.create(temp);
-
+    let user = await User.create({ name, email, password });
     return res.status(201).json(user);
   } catch (error) {
-    return res.status(501).json(error);
+    console.error("Error adding user:", error);
+    return res.status(500).json({ message: "Error adding user", error });
   }
 };
 
 exports.updateUser = async (req, res) => {
   const id = req.params.id;
-  const temp = {
-    name: req.body.name,
-    email: req.body.email,
-    password: req.body.password,
-  };
+  const { name, email, password } = req.body;
 
   try {
-    let user = await User.findById({ _id: id });
+    let user = await User.findById(id);
 
-    if (user) {
-      Object.keys(temp).forEach((key) => {
-        if (!!temp[key]) {
-          user[key] = temp[key];
-        }
-      });
-
-      await user.save();
-      return res.status(201).json(user);
+    if (!user) {
+      return res.status(404).json("Utilisateur non trouvé");
     }
 
-    return res.status(404).json("Utilisateur non trouvé");
+    // Mettre à jour les propriétés non vides de l'utilisateur
+    if (name) user.name = name;
+    if (email) user.email = email;
+    if (password) user.password = password;
+
+    await user.save();
+    return res.status(200).json(user);
   } catch (error) {
-    return res.status(501).json(error);
+    console.error("Error updating user:", error);
+    return res.status(500).json({ message: "Error updating user", error });
   }
 };
 
@@ -49,11 +41,17 @@ exports.deleteUser = async (req, res) => {
   const id = req.params.id;
 
   try {
-    await User.findByIdAndDelete(id);
+    let user = await User.findById(id);
 
+    if (!user) {
+      return res.status(404).json("Utilisateur non trouvé");
+    }
+
+    await User.findByIdAndDelete(id);
     return res.status(204).json("Utilisateur supprimé");
   } catch (error) {
-    return res.status(501).json(error);
+    console.error("Error deleting user:", error);
+    return res.status(500).json({ message: "Error deleting user", error });
   }
 };
 
@@ -67,19 +65,26 @@ exports.authenticate = async (req, res) => {
       return res.status(404).json({ message: "Utilisateur non trouvé" });
     }
 
-    if (user.password !== password) {
+    const isMatch = await user.comparePassword(password);
+
+    if (!isMatch) {
       return res.status(401).json({ message: "Mot de passe incorrect" });
     }
 
+    console.log('Secret Key:', process.env.SECRET_KEY);
     const token = jwt.sign({ userId: user._id }, process.env.SECRET_KEY, {
       expiresIn: "1d",
     });
 
+    console.log('Generated token:', token);
+
+    // Utiliser un cookie
+    res.cookie('token', token, { httpOnly: true, secure: true });
+
     res.json({ success: true, user, token });
   } catch (error) {
     console.error("Erreur d'authentification :", error);
-    res
-      .status(500)
-      .json({ message: "Erreur de serveur lors de l'authentification" });
+    res.status(500).json({ message: "Erreur de serveur lors de l'authentification" });
   }
 };
+
